Cart api (#630)
Use Shift's cart API instead of storing the cart in the local storage.

What changed?
 ======================

This commit completely rewrites the cart so that it is stored in Shift and any changes to the cart and line items are made using Shift's cart API. The cart is no longer stored in local storage.

Why are we doing this?
 ======================

Storing the cart in the local storage was only a temporary measure to get the development of the reference site started. Storing carts in Shift is necessary so that we can use promotions and do things like removing the cart from the user's browser when they log out and retrieving it in when they log back in.