Cart api coupons (#665)
Adds a form for submitting coupons

What changed?
==================
This commit adds a new CouponForm component which can be used for submitting discount coupons. The components forces the Redux cart state to be refreshed on successful submissions and displays errors to the user otherwise. It is used on cart and checkout pages.

Why are we doing this?
==================
We are adding the new component so that customers can use coupon codes.