Extract getAccount express handler into shift-next module (#681)
Creates shift-next module and extracts getAccount Express handler into it.

What changed?
===============
A new module, shift-next, has been created in the vendors/ directory and added to the project as a yarn workspace. The Express handler to fetch user accounts from Shift has been extracted into the new module together with specs and any other needed files.

Why are we doing is?
===============
This PR is the initial PR and the proof of concept for building the shift-next package. Shift-next package will eventually be a stand-alone Node module our clients will be able to use to quickly bootstrap a Next.js based ecommerce site. It will include Express handlers and Next.js pages.